# packages/rest-api/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Synapse Protocol REST API",
    "version": "1.7.0",
    "description": "API documentation for the Synapse Protocol REST API"
  },
  "servers": [
    {
      "url": "https://api.synapseprotocol.com",
      "description": "Production server"
    },
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    }
  ],
  "paths": {
...
    "/swap": {
      "get": {
        "summary": "Get swap quote for tokens on a specific chain",
        "description": "Retrieve detailed swap quote for exchanging one token for another on a specified chain",
        "parameters": [
          {
            "in": "query",
            "name": "chain",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "The chain ID where the swap will occur"
          },
          {
            "in": "query",
            "name": "fromToken",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The address of the token to swap from"
          },
          {
            "in": "query",
            "name": "toToken",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The address of the token to swap to"
          },
          {
            "in": "query",
            "name": "amount",
            "required": true,
            "schema": {
              "type": "number"
            },
            "description": "The amount of tokens to swap"
          },
          {
            "in": "query",
            "name": "address",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The address of the user"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "routerAddress": {
                      "type": "string",
                      "description": "The address of the router contract"
                    },
                    "maxAmountOut": {
                      "type": "string",
                      "description": "The maximum amount of tokens that will be received"
                    },
                    "query": {
                      "type": "object",
                      "properties": {
                        "swapAdapter": {
                          "type": "string",
                          "description": "The address of the swap adapter"
                        },
                        "tokenOut": {
                          "type": "string",
                          "description": "The address of the token being received"
                        },
                        "minAmountOut": {
                          "$ref": "#/components/schemas/BigNumber"
                        },
                        "deadline": {
                          "$ref": "#/components/schemas/BigNumber"
                        },
                        "rawParams": {
                          "type": "string",
                          "description": "Raw parameters for the swap"
                        }
                      }
                    }
                  }
                },
                "example": {
                  "routerAddress": "0x7E7A0e201FD38d3ADAA9523Da6C109a07118C96a",
                  "maxAmountOut": "999.746386",
                  "query": {
                    "swapAdapter": "0x7E7A0e201FD38d3ADAA9523Da6C109a07118C96a",
                    "tokenOut": "0xdAC17F958D2ee523a2206206994597C13D831ec7",
                    "minAmountOut": {
                      "type": "BigNumber",
                      "hex": "0x3b96eb52"
                    },
                    "deadline": {
                      "type": "BigNumber",
                      "hex": "0x66ecb470"
                    },
                    "rawParams": "0x00000000000000000000000000000000000000000000000000000000000000000000000000000000000000001116898dda4015ed8ddefb84b6e8bc24528af2d800000000000000000000000000000000000000000000000000000000000000010000000000000000000000000000000000000000000000000000000000000002"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "value": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        },
                        "field": {
                          "type": "string"
                        },
                        "location": {
                          "type": "string"
                        }
                      }
                    }
                  }
                },
                "example": {
                  "error": {
                    "value": "999",
                    "message": "Unsupported chain",
                    "field": "chain",
                    "location": "query"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": ["Synapse Bridge REST API"]
      }
    },
    "/swap/v2": {
      "get": {
        "summary": "Get swap quote for arbitrary tokens on a specific chain",
        "description": "Retrieve detailed swap quote exchanging any token for any other on a specified chain. Note that amounts are expressed in token's native decimals, not in human-readable format.",
        "parameters": [
          {
            "in": "query",
            "name": "chainId",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "The chain ID where the swap will occur"
          },
          {
            "in": "query",
            "name": "fromToken",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The address of the token to swap from"
          },
          {
            "in": "query",
            "name": "fromAmount",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The amount of tokens to swap in the token's native decimals"
          },
          {
            "in": "query",
            "name": "toToken",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The address of the token to swap to"
          },
          {
            "in": "query",
            "name": "toRecipient",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Optional. The address that will receive the swapped tokens. If provided, returns transaction data."
          },
          {
            "in": "query",
            "name": "slippage",
            "required": false,
            "schema": {
              "type": "number"
            },
            "description": "Optional. The maximum allowed slippage percentage (0-100). Defaults to 0.5%."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "description": "Unique identifier for the quote"
                    },
                    "chainId": {
                      "type": "integer",
                      "description": "The chain ID where the swap occurs"
                    },
                    "fromToken": {
                      "type": "string",
                      "description": "The address of the token being swapped from"
                    },
                    "fromAmount": {
                      "type": "string",
                      "description": "The amount of tokens to swap in the token's native decimals"
                    },
                    "toToken": {
                      "type": "string",
                      "description": "The address of the token being swapped to"
                    },
                    "expectedToAmount": {
                      "type": "string",
                      "description": "The expected amount of tokens that will be received in the token's native decimals"
                    },
                    "minToAmount": {
                      "type": "string",
                      "description": "The minimum amount of tokens that will be received in the token's native decimals (includes slippage)"
                    },
                    "routerAddress": {
                      "type": "string",
                      "description": "The address of the router contract"
                    },
                    "moduleNames": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      },
                      "description": "The names of the swap modules used for this quote"
                    },
                    "callData": {
                      "type": "object",
                      "nullable": true,
                      "description": "Transaction data object, only provided if toRecipient parameter was included",
                      "properties": {
                        "to": {
                          "type": "string",
                          "description": "Contract address to call"
                        },
                        "data": {
                          "type": "string",
                          "description": "Transaction calldata"
                        },
                        "value": {
                          "type": "string",
                          "description": "Amount of native currency to send with transaction (in native token decimals)"
                        }
                      }
                    }
                  }
                },
                "example": {
                  "id": "01920c87-7f14-7cdf-90e1-e13b2d4af55f",
                  "chainId": 1,
                  "fromToken": "0xa0b86991c6218b36c1d19d4a2e9eb0ce3606eb48",
                  "fromAmount": "1000000",
                  "toToken": "0xdac17f958d2ee523a2206206994597c13d831ec7",
                  "expectedToAmount": "999746386",
                  "minToAmount": "994747654",
                  "routerAddress": "0x512000a034E154908Efb1eC48579F4ffDb000512",
                  "moduleNames": ["DefaultPools"],
                  "callData": {
                    "to": "0x512000a034E154908Efb1eC48579F4ffDb000512",
                    "data": "0x...",
                    "value": "0"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "errors": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "value": {
                            "type": "string"
                          },
                          "msg": {
                            "type": "string"
                          },
                          "param": {
                            "type": "string"
                          },
                          "location": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                },
                "example": {
                  "errors": [
                    {
                      "value": "111",
                      "msg": "Unsupported chain",
                      "param": "chain",
                      "location": "query"
                    }
                  ]
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                },
                "example": {
                  "error": "An unexpected error occurred in /swap/v2. Please try again later."
                }
              }
            }
          }
        },
        "tags": ["Synapse Bridge REST API"]
      }
    },
...
  },
  "components": {
    "schemas": {
      "BigNumber": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": ["BigNumber"]
          },
          "hex": {
            "type": "string"
          }
        }
      },
...
    }
  },
...
}